requirements.txt
Django==5.2.5
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
PyJWT==2.10.1
psycopg2-binary==2.9.10
ua-parser==1.0.1
reportlab==4.4.3